Remove old code and refactor the rest
