--- src/main/java/com/drapson/springauthtutorial/adapters/in/api/request/LinkLocalAccountRequest.java
@@ -1,12 +1,0 @@
-package com.drapson.springauthtutorial.adapters.in.api.request;
-
-import jakarta.validation.constraints.NotNull;
-
-public record LinkLocalAccountRequest(
-        @NotNull
-        String linkToken,
-
-        @NotNull
-        boolean shouldLinkAccounts
-) {
-}
