--- src/main/java/com/drapson/springauthtutorial/adapters/in/security/AdditionalRegistrationInfoNeededException.java
@@ -1,15 +1,0 @@
-package com.drapson.springauthtutorial.adapters.in.security;
-
-public class AdditionalRegistrationInfoNeededException extends RuntimeException {
-
-    private final String registrationToken;
-
-    public AdditionalRegistrationInfoNeededException(String message, String registrationToken) {
-        super(message);
-        this.registrationToken = registrationToken;
-    }
-
-    public String getRegistrationToken() {
-        return registrationToken;
-    }
-}
