+++ src/main/java/com/drapson/springauthtutorial/adapters/in/security/JwtAuthenticationFilter.java
@@ -1,11 +1,12 @@
 package com.drapson.springauthtutorial.adapters.in.security;
 
-import com.drapson.springauthtutorial.application.TokenProvider;
+import com.drapson.springauthtutorial.application.out.TokenProvider;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
+import org.springframework.lang.NonNull;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.filter.OncePerRequestFilter;
@@ -21,7 +22,7 @@ public class JwtAuthenticationFilter extends OncePerRequestFilter {
     }
 
     @Override
-    protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain)
+    protected void doFilterInternal(@NonNull HttpServletRequest request, @NonNull HttpServletResponse response, @NonNull FilterChain filterChain)
             throws ServletException, IOException {
         String token = extractToken(request);
         if (token != null && tokenProvider.validateToken(token)) {
