--- src/main/java/com/drapson/springauthtutorial/application/TempUserDataPort.java
@@ -1,10 +1,0 @@
-package com.drapson.springauthtutorial.application;
-
-import java.time.Duration;
-import java.util.concurrent.TimeUnit;
-
-public interface TempUserDataPort {
-    void save(String key, Object data, Duration timeout);
-    Object get(String key);
-    void delete(String key);
-}
