+++ src/main/java/com/drapson/springauthtutorial/application/dtos/AuthTokens.java
@@ -1,7 +1,5 @@
 package com.drapson.springauthtutorial.application.dtos;
 
-import java.util.UUID;
-
 public record AuthTokens(
     String accessToken,
     String refreshToken
