--- src/main/java/com/drapson/springauthtutorial/application/dtos/CreatedUser.java
@@ -1,10 +1,0 @@
-package com.drapson.springauthtutorial.application.dtos;
-
-import java.util.UUID;
-
-public record CreatedUser(
-        UUID id,
-        String email,
-        String username
-) {
-}
