--- src/main/java/com/drapson/springauthtutorial/application/dtos/FinishOAuthRegistrationDto.java
@@ -1,12 +1,0 @@
-package com.drapson.springauthtutorial.application.dtos;
-
-import java.time.LocalDate;
-
-public record FinishOAuthRegistrationDto(
-        String token,
-        String username,
-        LocalDate birthDate,
-        boolean sendBudgetReports,
-        boolean isProfilePublic
-) {
-}
