--- src/main/java/com/drapson/springauthtutorial/application/dtos/LinkLocalAccountDto.java
@@ -1,7 +1,0 @@
-package com.drapson.springauthtutorial.application.dtos;
-
-public record LinkLocalAccountDto(
-        String linkToken,
-        boolean shouldLinkAccounts
-) {
-}
