+++ src/main/java/com/drapson/springauthtutorial/application/dtos/LinkOAuthAccountDto.java
@@ -1,7 +1,5 @@
 package com.drapson.springauthtutorial.application.dtos;
 
-import jakarta.validation.constraints.NotNull;
-
 import java.util.UUID;
 
 public record LinkOAuthAccountDto(
