--- src/main/java/com/drapson/springauthtutorial/application/dtos/PendingLocalRegistration.java
@@ -1,9 +1,0 @@
-package com.drapson.springauthtutorial.application.dtos;
-
-import java.io.Serializable;
-
-public record PendingLocalRegistration(
-        String email,
-        String hashedPassword
-) implements Serializable {
-}
