+++ src/main/java/com/drapson/springauthtutorial/config/ServiceConfiguration.java
@@ -1,13 +1,18 @@
 package com.drapson.springauthtutorial.config;
 
 import com.drapson.springauthtutorial.adapters.in.security.OAuth2CodeServiceImpl;
+import com.drapson.springauthtutorial.adapters.in.security.UserDetailsServiceImpl;
 import com.drapson.springauthtutorial.application.*;
+import com.drapson.springauthtutorial.application.out.OAuth2CodeService;
+import com.drapson.springauthtutorial.application.out.RefreshTokenRepository;
+import com.drapson.springauthtutorial.application.out.TokenProvider;
+import com.drapson.springauthtutorial.application.out.UserProviderRepository;
 import com.drapson.springauthtutorial.domain.UserRepository;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.transaction.support.TransactionTemplate;
 import org.springframework.web.reactive.function.client.WebClient;
 
 @Configuration
@@ -24,16 +29,14 @@ public class ServiceConfiguration {
             RefreshTokenRepository refreshTokenRepository,
             UserProviderRepository userProviderRepository,
             BCryptPasswordEncoder passwordEncoder,
-            TokenProvider tokenProvider,
-            TempUserDataPort tempUserDataPort
+            TokenProvider tokenProvider
     ) {
         return new AuthServiceImpl(
                 userRepository,
                 refreshTokenRepository,
                 userProviderRepository,
                 passwordEncoder,
-                tokenProvider,
-                tempUserDataPort
+                tokenProvider
         );
     }
 
@@ -59,6 +62,11 @@ public class ServiceConfiguration {
             @Value("${spring.security.oauth2.client.registration.google.client-secret}") String googleClientSecret
     ) {
         return new OAuth2CodeServiceImpl(webClient, googleClientId, googleClientSecret);
+    }
+
+    @Bean
+    public UserDetailsService userDetailsService(UserRepository userRepository) {
+        return new UserDetailsServiceImpl(userRepository);
     }
 
 }
